Moved errors module to utils
